KTOR-4495 Ignore server ContentNegotiation converter exceptions
